Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "chromadb>=1.5.1",
    "email-validator>=2.3.0",
    "fastapi>=0.131.0",
    "google-genai>=1.64.0",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "python-socketio>=5.16.1",
    "uvicorn>=0.41.0",
]
